Generate random image from file
