+++ meme-generator/memesData.js
@@ -3,25 +3,25 @@ export default {
     "data": {
         "memes": [
             {
-                "id": "1",
-                "name": "meme name 1",
-                "url": "meme url 1",
+                "id": "2",
+                "name": "Giraffe",
+                "url": "https://upload.wikimedia.org/wikipedia/commons/thumb/9/9e/Giraffe_Mikumi_National_Park.jpg/800px-Giraffe_Mikumi_National_Park.jpg",
+                "width": 1200,
+                "height": 1200,
+                "box_count": 2,
+            },
+            {
+                "id": "2",
+                "name": "Whale",
+                "url": "https://upload.wikimedia.org/wikipedia/commons/thumb/e/e5/Anim1754_-_Flickr_-_NOAA_Photo_Library_%281%29.jpg/800px-Anim1754_-_Flickr_-_NOAA_Photo_Library_%281%29.jpg",
                 "width": 1200,
                 "height": 1200,
                 "box_count": 2,
             },
             {
                 "id": "3",
-                "name": "meme name 2",
-                "url": "meme url 2",
-                "width": 1200,
-                "height": 1200,
-                "box_count": 2,
-            },
-            {
-                "id": "3",
-                "name": "meme name 3",
-                "url": "meme url 3",
+                "name": "Bird",
+                "url": "https://upload.wikimedia.org/wikipedia/commons/thumb/2/26/Ultramarine_Flycatcher_%28Ficedula_superciliaris%29_Naggar%2C_Himachal_Pradesh%2C_2013_%28cropped%29.JPG/800px-Ultramarine_Flycatcher_%28Ficedula_superciliaris%29_Naggar%2C_Himachal_Pradesh%2C_2013_%28cropped%29.JPG",
                 "width": 1200,
                 "height": 1200,
                 "box_count": 2,
