+++ meme-generator/src/scss/_meme.scss
@@ -24,3 +24,9 @@
     @include round-wide-input;
     border-width: 0px;
 }
+
+.memage {
+    margin-top: 15px;
+    max-width:100%;
+    border-radius: 5px;
+}
