+++ meme-generator/src/scss/main.scss
@@ -3,7 +3,7 @@ $bright-font-colour: #FEFEFE;
 $light-gray: #DDDDDD;
 
 @import "mixins";
-@import "controls";
+@import "meme";
 @import "header";
 
 * {
